more general cleanup, now going to incorporate ranges

+++ src/trans/trans_util/test/bessel_K1.test.cpp
@@ -14,7 +14,7 @@ TEST_CASE( "bessel_K1 generating functin" ){
         std::vector<double> correct { 199.98521432, 99.9738941, 9.85384478, 
                                         1.05028353, 0.60293127, 0.6019072 }; 
         THEN( "the modified bessel function value is returned" ){
-            for( int i = 0; i < x_vec.size(); ++i ){
+            for( size_t i = 0; i < x_vec.size(); ++i ){
                 equal( bessel_K1_gen( x_vec[i] ), correct[i] );
             } // for
         } // THEN
@@ -24,7 +24,7 @@ TEST_CASE( "bessel_K1 generating functin" ){
         std::vector<double> correct {1.6304576457, 1.2431658735, 0.6002738602, 
                                      0.4107665699, 0.1257999504}; 
         THEN( "the modified bessel function value is returned" ){
-            for( int i = 0; i < x_vec.size(); ++i ){
+            for( size_t i = 0; i < x_vec.size(); ++i ){
                 equal( bessel_K1_gen( x_vec[i] ), correct[i] );
             } // for
         } // THEN
